Change addon summary to something smaller
Dear jmarth,

The skin I'm using is aeon nox and the font size is huge. Kmediatorrent brings issues due to the fact of having so many text on the addon summary. It basically overlays the text on top of other addons on the list. Please consider to rename the addon summary to something smaller to improve compatibility with other skins.

Thanks

PS: My change is just an example